requirements.txt
Flask==2.2.2
google-api-python-client==2.75.0
requests==2.28.1
pydub==0.25.1         # For audio-to-text conversion
speechrecognition==3.8.1  # For converting speech to text
python-dotenv==0.20.0
werkzeug==2.3.7
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0
PyPDF2==3.0.1
google-generativeai==0.3.2
pymongo==4.6.1